+++ daeng/src/firebase/firebase.js
@@ -1,8 +1,6 @@
-// firebase.js
 import { initializeApp } from "firebase/app";
 import { getMessaging, onMessage } from 'firebase/messaging';
 
-// Firebase 설정
 const firebaseConfig = {
   apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
   authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
@@ -12,24 +10,17 @@ const firebaseConfig = {
   appId: import.meta.env.VITE_FIREBASE_APP_ID,
 };
 
-// Firebase 앱 초기화
 const app = initializeApp(firebaseConfig);
-
-// Firebase Messaging 초기화
 const messaging = getMessaging(app);
 
-// 포그라운드에서 푸시 알림을 수신하고 처리하는 핸들러
 onMessage(messaging, (payload) => {
   console.log("푸시 알림 받음: ", payload);
 
-  // 알림 정보 추출
   const notificationTitle = payload.notification.title;
   const notificationOptions = {
     body: payload.notification.body,
-    icon: payload.notification.icon, // 알림 아이콘 (선택 사항)
+    icon: payload.notification.icon, 
   };
-
-  // 알림을 화면에 표시 (브라우저 알림으로)
   new Notification(notificationTitle, notificationOptions);
 });
