+++ daeng/src/pages/alarm/AlarmPage.jsx
@@ -4,7 +4,6 @@ import AlarmBox from '../../components/alarm/AlarmBox';
 import AlarmButton from '../../components/alarm/AlarmButton';
 import DivisionLine from '../../components/alarm/DivisionLine';
 import Footer from '../../components/commons/Footer';
-import TestButton from './TestBtn';
 
 function AlarmPage() {
   useEffect(() => {
@@ -31,7 +30,6 @@ function AlarmPage() {
       <Header label="알림" />
       <AlarmBox />
       <AlarmButton />
-      <TestButton />
       <DivisionLine />
       <Footer />
     </>
